requirements.txt
GitPython>=3.1.40
requests>=2.31.0
